use usize for indexes

+++ src/actions/remove.rs
@@ -4,11 +4,11 @@ use anyhow::{Error, Ok};
 
 use crate::data::{self, write_tasks};
 
-pub fn remove(data_file: &PathBuf, index: u8) -> Result<(), Error> {
+pub fn remove(data_file: &PathBuf, index: usize) -> Result<(), Error> {
     let mut tasks = data::read_tasks(data_file)?;
-    match tasks.get(index as usize) {
+    match tasks.get(index) {
         Some(_) => {
-            tasks.remove(index as usize);
+            tasks.remove(index);
         }
         None => {
             return Err(Error::msg("no task found at provided index"));
